backend/requirements.txt
Django>=4.2
djangorestframework
django-environ
psycopg[binary]
gunicorn
